fix(examples): fix client bug and add examples
fix client bug and add examples
